requirements.txt
neat-python==0.92
pygame>=2.5.2
numpy>=2.1.0
matplotlib>=3.9.0
imageio>=2.34.0
imageio-ffmpeg>=0.5.0
pyvirtualdisplay==3.0
